make iz presence roaming too, presence refactors
